backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
cachetools==5.3.2
pydantic==2.5.0
slowapi==0.1.9
